Move timemachine to pretix.base, revert package refactor
