+++ src/pretix/base/models/items.py
@@ -61,11 +61,11 @@ from django_countries.fields import Country
 from django_scopes import ScopedManager
 from i18nfield.fields import I18nCharField, I18nTextField
 
-from pretix.base.middleware.timemachine import time_machine_now
 from pretix.base.models import fields
 from pretix.base.models.base import LoggedModel
 from pretix.base.models.fields import MultiStringField
 from pretix.base.models.tax import TaxedPrice
+from pretix.base.timemachine import time_machine_now
 
 from ...helpers.images import ImageSizeValidator
 from ..media import MEDIA_TYPES
